Prototype rewrite of Argo
This is a fundamental break from the React-style component model Argo
was originally based on.

Components have been replaced with functions that behave more like view
constructors (still trying to think of an appropriate name).

The are used once to render a DOM subtree and bind its specific
properties to the view model, but once rendered, the view properties
are mutated directly according to the the binding rules.

This is a first pass at the code.  Not all cases are covered. Expect
further iterations on the implementation.